moved main panel to its own component
